requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
sqlalchemy>=2.0.28
python-dotenv>=1.0.1
google-genai>=1.63.0


pymongo>=4.6.3
python-docx>=1.1.0
pdfplumber>=0.10.4
sentence-transformers>=2.6.0
